refactor(cli): split src/cli.ts for 450-line rule
