+++ src/cli/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * CLI 공통 상수
+ */
+
+export const DEFAULT_BASE_URL = 'https://mcp.aka.page';
+export const DEFAULT_MCP_URL = `${DEFAULT_BASE_URL}/mcp`;
